ADD: first working version of monster projection

+++ bonus/src/raycasting/dda_helper_bonus.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   dda_helper_bonus.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: boenkhja <boenkhja@student.42vienna.com>   +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/05/05 18:49:20 by boenkhja          #+#    #+#             */
+/*   Updated: 2026/05/07 15:25:53 by boenkhja         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../../includes/game_bonus.h"
 #include <math.h>
 #include "../../../libraries/libft/libft.h"
 
-void	cast_ray(t_raycast *ray, t_game *game, int x)
+void	cast_ray(t_raycast *ray, t_game *game, size_t x)
 {
 	ft_memset(ray, 0, sizeof(t_raycast));
 	calculate_ray_dir(ray, &game->player, x);
@@ -16,18 +28,19 @@ void	cast_ray(t_raycast *ray, t_game *game, int x)
 void	calculate_wall_height(t_raycast *ray, t_game *game)
 {
 	if (ray->side == 0)
-    	ray->perp_dist = (ray->map_x - game->player.pos.x + (1 - ray->step_x) / 2.0) / ray->ray_dir.x;
+		ray->perp_dist = (ray->map_x - game->player.pos.x
+				+ (1 - ray->step_x) / 2.0) / ray->ray_dir.x;
 	else
-    	ray->perp_dist = (ray->map_y - game->player.pos.y + (1 - ray->step_y) / 2.0) / ray->ray_dir.y;
+		ray->perp_dist = (ray->map_y - game->player.pos.y
+				+ (1 - ray->step_y) / 2.0) / ray->ray_dir.y;
 	if (ray->perp_dist < 0.0001)
 		ray->perp_dist = 0.0001;
 	ray->wall_height = (HEIGHT / ray->perp_dist);
 	if (ray->side == 0)
 		ray->wall_hit = game->player.pos.y + ray->perp_dist * ray->ray_dir.y;
 	else
-	 	ray->wall_hit = game->player.pos.x + ray->perp_dist * ray->ray_dir.x;
+		ray->wall_hit = game->player.pos.x + ray->perp_dist * ray->ray_dir.x;
 	ray->wall_hit -= floor(ray->wall_hit);
-	//printf("perp dist: %f wall_height: %d\n", ray->perp_dist, ray->wall_height);
 }
 
 void	init_step_dir(t_raycast *ray, t_vec	ray_dir)
@@ -35,9 +48,9 @@ void	init_step_dir(t_raycast *ray, t_vec	ray_dir)
 	if (ray_dir.x < 0)
 		ray->step_x = -1;
 	else
-	 	ray->step_x = 1;
+		ray->step_x = 1;
 	if (ray_dir.y < 0)
 		ray->step_y = -1;
 	else
-	 	ray->step_y = 1;
+		ray->step_y = 1;
 }
